# backend/docs/tourplanner-openapi.yaml
openapi: 3.0.3
info:
  title: Tour Planner API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/users/me/favorites/{tourId}:
    post:
      summary: Add tour to favorites
      security: [ { bearerAuth: [] } ]
      parameters:
        - $ref: '#/components/parameters/TourId'
      responses:
        '204':
          description: Tour successfully added to favorites

    delete:
      summary: Remove tour from favorites
      security: [ { bearerAuth: [] } ]
      parameters:
        - $ref: '#/components/parameters/TourId'
      responses:
        '204':
          description: Tour successfully removed from favorites
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TourId:
      name: tourId
      in: path
      required: true
      description: UUID of the tour
      schema:
        type: string
        format: uuid
...